Moved the definition of `struct ID` in the DNA module
