+++ source/CMakeLists.txt
@@ -2,6 +2,7 @@
 # Declare DNA Files
 
 set(DNA_TYPEDEF_HEADERS
+	${CMAKE_SOURCE_DIR}/source/rose/makedna/DNA_ID.h
 	${CMAKE_SOURCE_DIR}/source/rose/makedna/DNA_listbase.h
 )
